requirements.txt
Flask
numpy 
torch
torchvision
Pillow